move Google logging to server
This moves the Google logging to the server.  I also have provided
sample config files for both client and server.

Fixed a minor issue where testing mode was not obvious enough.  More
fixes to come on this.
