+++ arcom.conf.sample
@@ -1,0 +1,8 @@
+; Config file for arcom.py
+; call should be your FCC call sign
+; location can be any short descriptive string to be included in the weblog entry.
+; Grid location is a good choice.  Or it could be descriptive like 'Home' depending
+; on how you plan to use the weblog data.
+[arcom]
+call = N1LXA
+location = FN34ld
